reuse logic, fix types
